+++ ExpandWorld/data/ClutterManager.cs
@@ -126,6 +126,7 @@ public class ClutterManager
     }
     catch (Exception e)
     {
+      ExpandWorld.Log.LogError(e.Message);
       ExpandWorld.Log.LogError(e.StackTrace);
     }
   }