+++ ExpandWorld/data/DungeonManager.cs
@@ -123,6 +123,7 @@ public class DungeonManager
     }
     catch (Exception e)
     {
+      ExpandWorld.Log.LogError(e.Message);
       ExpandWorld.Log.LogError(e.StackTrace);
     }
   }