+++ ExpandWorld/data/EventManager.cs
@@ -81,6 +81,7 @@ public class EventManager
     }
     catch (Exception e)
     {
+      ExpandWorld.Log.LogError(e.Message);
       ExpandWorld.Log.LogError(e.StackTrace);
     }
   }