+++ ExpandWorld/data/NoBuildManager.cs
@@ -77,6 +77,7 @@ public class NoBuildManager
     }
     catch (Exception e)
     {
+      ExpandWorld.Log.LogError(e.Message);
       ExpandWorld.Log.LogError(e.StackTrace);
     }
   }