+++ ExpandWorld/data/SpawnManager.cs
@@ -159,6 +159,7 @@ public class SpawnManager
     }
     catch (Exception e)
     {
+      ExpandWorld.Log.LogError(e.Message);
       ExpandWorld.Log.LogError(e.StackTrace);
     }
   }