+++ ExpandWorld/data/WorldManager.cs
@@ -140,6 +140,7 @@ public class WorldManager
     }
     catch (Exception e)
     {
+      ExpandWorld.Log.LogError(e.Message);
       ExpandWorld.Log.LogError(e.StackTrace);
     }
   }