+++ publish/CHANGELOG.md
@@ -1,5 +1,6 @@
 - v1.39
   - Fixes corrupted or missing locations possibly causing an error during the gameplay.
+  - Fixes dungeon room variants not working.
 
 - v1.38
   - Fixes invalid vegetation not being removed from the game system.